Enables testing with or without sync.Map for globalEvidenceSealedMap
This commit introduces a custom thread safe Map, which exposes the same API as sync.Map but has a different implementation; the idea is to be able to do benchmarking from this, so I added a flag for the start command, --useSyncMap, which defaults to true. Setting it to false allows running the node with the custom map instead.